Move crownConfig into directory
